fix:parseInt to ParseFloat

+++ 1-lesson/first-task/app.js
@@ -51,7 +51,7 @@ function isSpace(value) {
 
 function isNumber(value) {
     if (value != '0') {
-        if (parseInt(value)) {
+        if (parseFloat(value)) {
             return true;
         } else {
             return false;
